Prompt user to delete everything in uninstaller

+++ YazDHCP.sh
@@ -771,7 +771,19 @@ Menu_Uninstall(){
 	Auto_DNSMASQ delete 2>/dev/null
 	Shortcut_script delete
 	umount /www/Advanced_DHCP_Content.asp 2>/dev/null
-	rm -rf "$SCRIPT_DIR" 2>/dev/null
+	rm -f "$SCRIPT_DIR/Advanced_DHCP_Content.asp"
+	
+	printf "\\n\\e[1mDo you want to delete %s DHCP client files? (y/n):    \\e[0m" "$SCRIPT_NAME"
+	read -r confirm
+	case "$confirm" in
+		y|Y)
+			rm -rf "$SCRIPT_DIR" 2>/dev/null
+		;;
+		*)
+			:
+		;;
+	esac
+	
 	rm -rf "$SCRIPT_WEB_DIR" 2>/dev/null
 	rm -f "/jffs/scripts/$SCRIPT_NAME" 2>/dev/null
 	Clear_Lock
